maschine: midi_note_base can be set via OSC API
This commit introduces the concept of a MIDI base note, to allow
applications to set the lowest MIDI note that will be sent for the pads.

This is useful if an application expects a specific "octave" to be
selected, and avoids users having to manually select it.

Documentation updated, section for midi_note_base added.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>